Increase size of inputs, outputs and witnesses to uint16

+++ fuel-tx/src/transaction/consensus_parameters.rs
@@ -227,9 +227,9 @@ impl Default for PredicateParameters {
 #[cfg_attr(feature = "serde", serde(default))]
 pub struct TxParameters {
     /// Maximum number of inputs.
-    pub max_inputs: u8,
+    pub max_inputs: u16,
     /// Maximum number of outputs.
-    pub max_outputs: u8,
+    pub max_outputs: u16,
     /// Maximum number of witnesses.
     pub max_witnesses: u32,
     /// Maximum gas per transaction.
@@ -257,13 +257,13 @@ impl TxParameters {
     }
 
     /// Replace the max inputs with the given argument
-    pub const fn with_max_inputs(mut self, max_inputs: u8) -> Self {
+    pub const fn with_max_inputs(mut self, max_inputs: u16) -> Self {
         self.max_inputs = max_inputs;
         self
     }
 
     /// Replace the max outputs with the given argument
-    pub const fn with_max_outputs(mut self, max_outputs: u8) -> Self {
+    pub const fn with_max_outputs(mut self, max_outputs: u16) -> Self {
         self.max_outputs = max_outputs;
         self
     }
@@ -400,8 +400,8 @@ pub mod default_parameters {
     use fuel_types::ChainId;
 
     pub const CONTRACT_MAX_SIZE: u64 = ContractParameters::DEFAULT.contract_max_size;
-    pub const MAX_INPUTS: u8 = TxParameters::DEFAULT.max_inputs;
-    pub const MAX_OUTPUTS: u8 = TxParameters::DEFAULT.max_outputs;
+    pub const MAX_INPUTS: u16 = TxParameters::DEFAULT.max_inputs;
+    pub const MAX_OUTPUTS: u16 = TxParameters::DEFAULT.max_outputs;
     pub const MAX_WITNESSES: u32 = TxParameters::DEFAULT.max_witnesses;
     pub const MAX_GAS_PER_TX: u64 = TxParameters::DEFAULT.max_gas_per_tx;
