+++ fuel-vm/src/interpreter.rs
@@ -142,7 +142,7 @@ pub struct InterpreterParams {
     /// Gas costs
     pub gas_costs: GasCosts,
     /// Maximum number of inputs
-    pub max_inputs: u8,
+    pub max_inputs: u16,
     /// Maximum size of the contract in bytes
     pub contract_max_size: u64,
     /// Offset of the transaction data in the memory
@@ -244,7 +244,7 @@ impl<S, Tx, Ecal> Interpreter<S, Tx, Ecal> {
     }
 
     /// Get max_inputs value
-    pub fn max_inputs(&self) -> u8 {
+    pub fn max_inputs(&self) -> u16 {
         self.interpreter_params.max_inputs
     }
